# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookify API Documentation",
        "description": "API documentation for Bookify application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get paginated list of items",
                "operationId": "5561acef6e4aef27354515aca3069b72",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Harry Potter"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "finance"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 10000
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 50000
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "price:desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Items fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemCollectionResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/items/{slug}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get a single item by slug",
                "operationId": "1cb5c1319b9351252ea50088b4ecb27b",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "harry-potter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            }
        },
        "api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get authenticated user's orders",
                "operationId": "4f617f816f32634763e46b43b0e3196c",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved orders"
                    },
                    "500": {
                        "description": "Failed to retrieve orders"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create order from cart",
                "operationId": "e032b6008e87019ee904853c7a70a9f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Sudirman No. 123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully from cart"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/v1/items/{productId}/order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create order directly from a product",
                "operationId": "bf807119588d0793f8dd4e01cf99a014",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/v1/orders/{orderId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get detail of a single authenticated user order",
                "operationId": "ffefad90cc26273c1e794421ee685433",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "order-12345"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details fetched successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "name": {
                        "type": "string",
                        "example": "Finance"
                    },
                    "slug": {
                        "type": "string",
                        "example": "finance"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Item": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "title": {
                        "type": "string",
                        "example": "Sample Book Title"
                    },
                    "slug": {
                        "type": "string",
                        "example": "sample-book"
                    },
                    "author": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "example": 120000
                    },
                    "stock": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ItemPaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ItemPaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string"
                    },
                    "last": {
                        "type": "string"
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ItemData": {
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Item"
                        }
                    },
                    "filters_used": {
                        "type": "object"
                    },
                    "links": {
                        "$ref": "#/components/schemas/ItemPaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/ItemPaginationMeta"
                    }
                },
                "type": "object"
            },
            "ItemCollectionResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Items fetched successfully"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ItemData"
                    }
                },
                "type": "object"
            },
            "ItemResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Item fetched successfully"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Item"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}